cuckoo-cycle

+++ src/ckpool.h
@@ -235,9 +235,9 @@ struct ckpool_instance {
 	int nonce2length; // Extranonce2 length
 
 	/* Difficulty settings */
-	int64_t mindiff; // Default 1
-	int64_t startdiff; // Default 42
-	int64_t maxdiff; // No default
+	double mindiff; // Default 1
+	double startdiff; // Default 42
+	double maxdiff; // No default
 
 	/* Coinbase data */
 	char *btcaddress; // Address to mine to
